// src/app/shared/model/update-master-cust/update-cust-personal-fin-data-obj.model.ts
export class UpdateCustPersonalFinDataObj {
    CustPersonalFinDataId: any;
    CustPersonalId: any;
    CustId: any;
    MonthlyIncomeAmt: any;
    MonthlyExpenseAmt: any;
    MonthlyInstallmentAmt: any;
    MrSourceOfIncomeCode: any;
    SpouseMonthlyIncomeAmt: any;
    IsJoinIncome: any;
    TotalIncomeAmt: any;
    NettIncomeAmt: any;
    CustBankAccList: Array<any>;
}